requirements.txt
django==5.0.2
python-dotenv==1.0.0
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
